Editing timer values to floats and yellow function
Errors in retracing movement were due to inconsistencies in value types and also the sequence of subfunctions in the card functions

+++ main.h
@@ -15,7 +15,7 @@
 volatile unsigned int maxTime = 0;//define max time between counters for an overflow 
 volatile unsigned int movementCount = 0;//define the movement counter
 volatile unsigned int movementMemory[20] = {};//corresponds to R90,L90,T180
-volatile unsigned int timerMemory[20] = {};//this corresponds to 1 second passing
+volatile float timerMemory[20] = {};//this corresponds to 1 second passing
 void main(void);
 
 #endif /* MAIN_H */
